Convert u64->u128, and not u128->u64
Recently, `From<u128> for StarkFelt` was implemented, so no need to cast
u128 integers (`Fee` is `Fee(pub u128)`) to u64 before constructing a
felt.

+++ crates/blockifier/src/transaction/account_transaction.rs
@@ -207,7 +207,7 @@ impl AccountTransaction {
         }
 
         // The least significant 128 bits of the amount transferred.
-        let lsb_amount = StarkFelt::from(actual_fee.0 as u64);
+        let lsb_amount = StarkFelt::from(actual_fee.0);
         // The most significant 128 bits of the amount transferred.
         let msb_amount = StarkFelt::from(0_u8);
